// src/domain.types/hotel.review/hotel.review.dto.ts
/* eslint-disable linebreak-style */
/* eslint-disable eol-last */
/* eslint-disable linebreak-style */

export interface HotelReviewDto {
    id?: string;
    HotelId: string;
    ReviewUserId: string;
    Rating: string;
    ReviewTitle: string;
    ReviewDescription: string;
    ReviewTimeStamp: Date;
}